more XML

+++ java/jDrumsEmuApplet/src/de/zorgk/drums/RestRiff.java
@@ -13,9 +13,9 @@ public class RestRiff implements RiffInterface, Cloneable {
 
 	private long samples_left;
 
-	public RestRiff(long beatsPerMinute, float beats) {
-		this.samples_left = (long) ((beats * RiffInterface.framesPerSecond * 60))
-				/ (beatsPerMinute);
+	public RestRiff(float beatsPerMinute, float beats) {
+		this.samples_left = (long) (((beats * RiffInterface.framesPerSecond * 60))
+				/ (beatsPerMinute));
 	}
 
 	@Override
